Modify UserRole dao to automatically assign permission
